feat: moving response function to the global variables and cleaning user route
