+++ example/CHANGELOG.md
@@ -1,1 +1,1 @@
-# Example
+# Example
