+++ example/README.md
@@ -1,3 +1,3 @@
-# Example
+# Example 
 
 This is a test.
