+++ example/src/utils/index.ts
@@ -1,4 +1,3 @@
 import docs from './docs.json'
-import { response } from './response'
 
-export { docs, response }
+export { docs }
