+++ lib/src/functions/express.js
@@ -58,7 +58,14 @@ import { DtoUser } from '../dto-interfaces'
 
 declare global {
   // This variable is global, so it will be available everywhere in the code
-  var myGlobalVariable: unknown
+  var response = (
+    error: boolean,
+    message: unknown,
+    res: Response,
+    status: number
+  ): void => {
+    res.status(status).send({ error, message })
+  }
 
   // We can personalize the response and request objects in case we need it by
   // adding new optional attributes to this interface
@@ -347,7 +354,7 @@ import swaggerUi from 'swagger-ui-express'
 import httpErrors from 'http-errors'
 
 import { Home, User } from '../routes'
-import { response, docs } from '../utils'
+import { docs } from '../utils'
 
 const routers = [User]
 
@@ -459,11 +466,11 @@ class Server {
       console.log(\`Server running at port \${PORT}\`)
     })
 
-    try {
-      this._mongo()
-    } catch (e) {
-      console.error(e)
-    }
+    // try {
+    // this._mongo()
+    // } catch (e) {
+    // console.error(e)
+    // }
   }
 }
 
@@ -485,7 +492,6 @@ export { applyRoutes, Server }
     routes: {
       home: {
         content: `import { Response, Request, Router } from 'express'
-import { response } from '../utils'
 
 const Home = Router()
 
@@ -499,7 +505,7 @@ export { Home }
       },
       index: {
         content: `import { Home } from './home'
-import { User } from './users'
+import { User } from './user'
 
 export { Home, User }
 `,
@@ -507,9 +513,9 @@ export { Home, User }
       },
       user: {
         content: `import { Router, NextFunction } from 'express'
+import httpErrors from 'http-errors'
+import { ValidationError } from 'joi'
 
-import { Response, Request } from '../custom'
-import { response } from '../utils'
 import { User as UserC } from '../controllers/user'
 import { DtoUser } from '../dto-interfaces'
 import { idSchema, userSchema } from '../schemas'
@@ -518,7 +524,11 @@ const User = Router()
 
 User.route('/users')
   .post(
-    async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+    async (
+      req: CustomRequest,
+      res: CustomResponse,
+      next: NextFunction
+    ): Promise<void> => {
       const {
         body: { args }
       } = req
@@ -533,7 +543,11 @@ User.route('/users')
     }
   )
   .get(
-    async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+    async (
+      req: CustomRequest,
+      res: CustomResponse,
+      next: NextFunction
+    ): Promise<void> => {
       const u = new UserC()
 
       try {
@@ -545,7 +559,11 @@ User.route('/users')
     }
   )
   .delete(
-    async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+    async (
+      req: CustomRequest,
+      res: CustomResponse,
+      next: NextFunction
+    ): Promise<void> => {
       const u = new UserC()
 
       try {
@@ -559,7 +577,11 @@ User.route('/users')
 
 User.route('/user/:id')
   .get(
-    async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+    async (
+      req: CustomRequest,
+      res: CustomResponse,
+      next: NextFunction
+    ): Promise<void> => {
       const {
         params: { id }
       } = req
@@ -570,13 +592,19 @@ User.route('/user/:id')
         const result = await u.process({ type: 'getOne' })
         response(false, result, res, 200)
       } catch (e) {
-        if (e.isJoi) e.status = 422
+        if (e instanceof ValidationError)
+          return next(new httpErrors.UnprocessableEntity(e.message))
+
         next(e)
       }
     }
   )
   .patch(
-    async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+    async (
+      req: CustomRequest,
+      res: CustomResponse,
+      next: NextFunction
+    ): Promise<void> => {
       const {
         body: { args },
         params: { id }
@@ -592,13 +620,19 @@ User.route('/user/:id')
         const result = await u.process({ type: 'update' })
         response(false, result, res, 200)
       } catch (e) {
-        if (e.isJoi) e.status = 422
+        if (e instanceof ValidationError)
+          return next(new httpErrors.UnprocessableEntity(e.message))
+
         next(e)
       }
     }
   )
   .delete(
-    async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+    async (
+      req: CustomRequest,
+      res: CustomResponse,
+      next: NextFunction
+    ): Promise<void> => {
       const {
         params: { id }
       } = req
@@ -609,7 +643,9 @@ User.route('/user/:id')
         const result = await u.process({ type: 'delete' })
         response(false, result, res, 200)
       } catch (e) {
-        if (e.isJoi) e.status = 422
+        if (e instanceof ValidationError)
+          return next(new httpErrors.UnprocessableEntity(e.message))
+
         next(e)
       }
     }
@@ -617,7 +653,7 @@ User.route('/user/:id')
 
 export { User }
 `,
-        file: `${projectName}/src/routes/users.ts`
+        file: `${projectName}/src/routes/user.ts`
       }
     },
     schemas: {
@@ -1108,9 +1144,8 @@ DELETE http://localhost:1996/api/user/60e7e3b93b01c1a7aa74cd6b
       },
       index: {
         content: `import docs from './docs.json'
-import { response } from './response'
 
-export { docs, response }
+export { docs }
 `,
         file: `${projectName}/src/utils/index.ts`
       },
